fix: correct divide() return type from int to double
The function computed a double result via an explicit cast but was
declared and defined to return int in both math.c and math.h, silently
truncating the value on return. The header's int prototype hid the
mismatch. Declaring divide() as double in both places matches the
README's documented signature and the intended floating-point result.

+++ sources/math.c
@@ -13,7 +13,7 @@ int multiply(int a, int b)
     return (a * b);
 }
 
-int divide(int a, int b)
+double divide(int a, int b)
 {
     if(b == 0)
         return (0);
